allow for larger bloom filters
Previously the maximum size of a bloom filter was a uint16.
Now it's a uint32.  This has no effect on query times.

+++ query.go
@@ -3,7 +3,7 @@
 package bloomindex
 
 //gc:nosplit
-func queryCore(r *bitrow, bits []bitrow, hashes []uint16) {
+func queryCore(r *bitrow, bits []bitrow, hashes []uint32) {
 
 	*r = bitrow{
 		0xffffffffffffffff, 0xffffffffffffffff, 0xffffffffffffffff, 0xffffffffffffffff,
