+++ query_stub.go
@@ -3,4 +3,4 @@
 package bloomindex
 
 //go:generate python -m peachpy.x86_64 query.py -S -o query_amd64.s -mabi=goasm
-func queryCore(r *bitrow, bits []bitrow, hashes []uint16)
+func queryCore(r *bitrow, bits []bitrow, hashes []uint32)
